inset date 온도 자료형 수정

+++ Entitys/WeatehrEntity.cs
@@ -8,10 +8,10 @@ namespace WeatherApplication.Models
     public class WeatherEntity
     {
         public string Region { get; set; }
-        public long Main_Temp { get; set; }
-        public long Min_Temp { get; set; }
-        public long Max_Temp { get; set; }
-        public long Feel_Temp { get; set; }
+        public double Main_Temp { get; set; }
+        public double Min_Temp { get; set; }
+        public double Max_Temp { get; set; }
+        public double Feel_Temp { get; set; }
         public long Wind { get; set; }
         public string Date_T { get; set; }
     }
